requirements.txt
aniso8601==3.0.2
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.8.24
chardet==3.0.4
click==6.7
coverage==4.5.1
Flask==1.0.2
Flask-JWT==0.3.2
Flask-JWT-Extended==3.13.0
Flask-RESTful==0.3.6
gunicorn==19.9.0
h11==0.7.0
idna==2.6
IMAPClient==2.0.0
itsdangerous==0.24
jdcal==1.4
Jinja2==2.10
jsonschema==2.6.0
lxml==4.2.5
MarkupSafe==1.0
more-itertools==4.3.0
openpyxl==2.1.4
Pillow==5.2.0
pluggy==0.7.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
py==1.6.0
pycrypto==2.6.1
PyJWT==1.4.2
PyMsgBox==1.0.6
PyPDF2==1.26.0
PyScreeze==0.1.14
PySocks==1.6.8
pytest==3.8.0
pytest-cov==2.6.0
pytest-flask==0.12.0
python-dateutil==2.7.3
python-docx==0.8.6
python-dotenv==0.9.1
python3-xlib==0.15
PyTweening==1.0.3
pytz==2018.5
requests==2.18.4
selenium==3.13.0
Send2Trash==1.5.0
six==1.11.0
twilio==6.14.7
urllib3==1.22
uWSGI==2.0.17.1
Werkzeug==0.14.1
